remaining plist and ptask functions moved to appropriate namespaces
